Moved layer editor scripts
