+++ Assets/_Application/Layers/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 07f884718b9fe492ba5049834e6f4f45
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
