+++ Assets/_Application/Layers/LayerTypes/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 59ea7249747c64777b2c4b03cda5133e
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
