+++ Assets/_Application/Layers/LayerTypes/GeoJsonLayers/Editor.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 3be3625c50b1d40ba82672ae05c32eb0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
