+++ Assets/_Application/Layers/LayerTypes/GeoJsonLayers/Editor/GeoJsonLayerGameObjectEditor.cs
@@ -1,13 +1,12 @@
-using Netherlands3D.Twin.Layers;
-using Netherlands3D.Twin.Layers.LayerTypes.GeoJsonLayers;
+using Netherlands3D.Twin.Layers.Editor;
 using UnityEditor;
 using UnityEditor.UIElements;
 using UnityEngine.UIElements;
 
-namespace Netherlands3D.Twin.Editor.Layers
+namespace Netherlands3D.Twin.Layers.LayerTypes.GeoJsonLayers.Editor
 {
-    [CustomEditor(typeof(GeoJSONPointLayer))]
-    public class GeoJsonPointLayerGameObjectEditor : UnityEditor.Editor
+    [CustomEditor(typeof(GeoJsonLayerGameObject))]
+    public class GeoJsonLayerGameObjectEditor : UnityEditor.Editor
     {
         public override VisualElement CreateInspectorGUI()
         {
@@ -15,7 +14,7 @@ namespace Netherlands3D.Twin.Editor.Layers
 
             InspectorElement.FillDefaultInspector(root, serializedObject, this);
 
-            var layerGameObject = (GeoJSONPointLayer)target;
+            var layerGameObject = (GeoJsonLayerGameObject)target;
             LayerDataVisualElements.LayerData(layerGameObject.LayerData, root);
 
             return root;
