+++ Assets/_Application/Layers/LayerTypes/GeoJsonLayers/Editor/GeoJsonLayerGameObjectEditor.cs.meta
