+++ Assets/_Application/Layers/LayerTypes/GeoJsonLayers/Editor/GeoJsonLineLayerGameObjectEditor.cs.meta
